@@ -24,27 +24,27 @@
   "config-overrides-path": "config/config-overrides",
   "dependencies": {
     "@adaptagency/fe_tools": "1.6.4",
-    "@types/jest": "^24.0.11",
-    "@types/node": "^11.13.4",
-    "@types/react": "^16.8.13",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-redux": "^7.0.6",
+    "@types/jest": "24.0.11",
+    "@types/node": "11.13.4",
+    "@types/react": "16.8.13",
+    "@types/react-dom": "16.8.4",
+    "@types/react-redux": "7.0.6",
     "node-sass": "4.9.4",
     "prop-types": "15.6.2",
-    "react": "16.6.0",
-    "react-dom": "16.6.0",
-    "react-loadable": "^5.5.0",
+    "react": "16.8.6",
+    "react-dom": "16.8.6",
+    "react-loadable": "5.5.0",
     "react-scripts": "2.1.0",
     "style-loader": "0.23.1",
-    "typescript": "^3.4.3"
+    "typescript": "3.4.3"
   },
   "devDependencies": {
-    "@hot-loader/react-dom": "^16.8.6",
+    "@hot-loader/react-dom": "16.8.6",
     "@svgr/webpack": "4.2.0",
     "autoprefixer": "9.5.1",
     "colors": "1.3.3",
     "css-loader": "2.1.1",
-    "customize-cra": "^0.2.12",
+    "customize-cra": "0.2.12",
     "eslint": "5.16.0",
     "eslint-config-airbnb": "17.1.0",
     "eslint-plugin-import": "2.16.0",
@@ -60,7 +60,7 @@
     "prompt": "1.0.0",
     "react-app-rewire-eslint": "0.2.3",
     "react-app-rewire-hot-loader": "2.0.1",
-    "react-app-rewire-typescript": "^2.0.2",
+    "react-app-rewire-typescript": "2.0.2",
     "react-app-rewired": "1.6.2",
     "react-hot-loader": "4.8.3",
     "release-it": "10.4.2",
